针对数据库修改数据类型

+++ ModManager/database/generate/Chunk.cs
@@ -8,7 +8,7 @@ namespace ModManager.database.generate;
 /// </summary>
 public partial class Chunk
 {
-    public int WorldId { get; set; }
+    public long WorldId { get; set; }
 
     /// <summary>
     /// X轴坐标
